make datasets into a package not module
