+++ bmds/datasets/__init__.py
@@ -1,0 +1,3 @@
+from .base import DatasetBase  # noqa
+from .continuous import ContinuousDataset, ContinuousIndividualDataset  # noqa
+from .dichotomous import DichotomousCancerDataset, DichotomousDataset  # noqa
