+++ bmds/datasets/anova.py
